move existing UserFactory
We'll use our factories to seed the DB
with fake data as well as testing
lets declare our factories in the app dir
instead of under /tests
