+++ project/conftest.py
@@ -2,7 +2,7 @@ import pytest
 from django.conf import settings
 from django.test import RequestFactory
 
-from users.tests.factories import UserFactory
+from users.factories import UserFactory
 
 
 @pytest.fixture(autouse=True)
